requirements.txt
streamlit==1.26.0
pillow
pandas
opencv-python
torch
ultralytics
